Version 1 (#1)
This is v1 of the integration with Akeneo. This version supports syncing an asset from Aprimo DAM to Akeneo and storing the public URL of the asset in Akeneo. In addition, text fields from Aprimo can be synced to the Akeneo asset and optionally the Akeneo asset can automatically be associated to a product.